Now all integer constants are u32.

+++ src/game/paddle.rs
@@ -6,8 +6,8 @@ use super::scene::SCREEN_WIDTH;
 
 use super::ball::DEFAULT_BALL_SPEED;
 
-pub const DEFAULT_PADDLE_HEIGHT: u16 = SCREEN_HEIGHT / 10;
-pub const DEFAULT_PADDLE_WIDTH: u16 = DEFAULT_PADDLE_HEIGHT / 8;
+pub const DEFAULT_PADDLE_HEIGHT: u32 = SCREEN_HEIGHT / 10;
+pub const DEFAULT_PADDLE_WIDTH: u32 = DEFAULT_PADDLE_HEIGHT / 8;
 pub const DEFAULT_PADDLE_SPEED: f32 = DEFAULT_BALL_SPEED;
 
 pub const DEFAULT_LEFT_CENTER_X: f32 = (PADDLE_MARGIN + DEFAULT_PADDLE_WIDTH / 2) as f32;
@@ -19,13 +19,13 @@ pub enum Sides {
 
 pub struct Paddle {
     pos: Vec2,
-    width: u16,
-    height: u16,
+    width: u32,
+    height: u32,
     side: Sides,
 }
 
 impl Paddle {
-    pub fn new(pos: Vec2, width: u16, height: u16, side: Sides) -> Self {
+    pub fn new(pos: Vec2, width: u32, height: u32, side: Sides) -> Self {
         Self {
             pos,
             width,
@@ -59,11 +59,11 @@ impl Paddle {
         &self.pos
     }
 
-    pub fn get_width(&self) -> u16 {
+    pub fn get_width(&self) -> u32 {
         self.width
     }
 
-    pub fn get_height(&self) -> u16 {
+    pub fn get_height(&self) -> u32 {
         self.height
     }
